+++ CHANGELOG.md
@@ -251,6 +251,8 @@ from_tbl: "foo".into_table_ref(),
 
 ### Minor breaking changes
 
+* Implemented `IntoTableRef` for `T: Into<TableRef>`. If you have manually implemented this trait, it may cause conflicts. You should rewrite it as `impl From<..> for TableRef`. https://github.com/SeaQL/sea-query/pull/958
+
 * Implemented `IntoCondition` for `T: Into<Cond>`. If you have manually implemented this trait, it may cause conflicts. You should rewrite it as `impl From<..> for Condition`.
 
 * Unboxed `Value` variants may cause compile error. Simply remove the `Box` in these cases https://github.com/SeaQL/sea-query/pull/925